+++ gingado/temporal/__init__.py
@@ -1,1 +1,1 @@
-from gingado.temporal.transformer import TemporalFeatureTransformer
+from gingado.temporal.features import get_timefeat, validate_and_get_freq
