+++ requirements.txt
@@ -5,3 +5,4 @@ scikit-learn>=1.0.2
 requests_cache>=0.9.1
 matplotlib>=3.5
 requests>=2.31.0
+keras>=3.5.0
